Style cleanups
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@779 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOSoundAudioSection.cpp
@@ -602,11 +602,11 @@ bool GOAudioSection::ReadBlock
 
 		/* Setup ptr and position required by the end-block */
 		stream->ptr              = stream->end_ptr;
-		stream->position_index  -= (float)stream->transition_position;
+		stream->position_index  -= stream->transition_position;
 		stream->end_decode_call(stream, buffer);
 
 		/* Restore the existing position */
-		stream->position_index  += (float)stream->transition_position;
+		stream->position_index  += stream->transition_position;
 		if (stream->position_index >= stream->section_length)
 		{
 			if (stream->next_start_segment_index < 0)
